analyzer/requirements.txt
fastapi
uvicorn[standard]
gitpython
requests
python-dotenv
pydantic
